+++ src/GroupElementCreationWrapper.tsx
@@ -7,10 +7,12 @@ import { Bounds, ElementCreationWrapperState, TargetType } from "./types/interna
 import { CustomButtonProps } from "./types/defaultComponents";
 import webGroupsManager from "./webGroupsManager";
 import webGroupsStore from "./webGroupsStore";
+import FrameLoadingAnimation from "./defaultComponents/loadingAnimation/FrameLoadingAnimation";
 
 const GroupElementCreationWrapper: React.FC<GroupProps> = ({ components }) => {
 
     const state = useSyncExternalStore<ElementCreationWrapperState>(webGroupsStore.subscribe, webGroupsStore.getSnapshot);
+    const LoadingAnimation =  components?.frame?.LoadingAnimation ?? FrameLoadingAnimation;
 
     const renderGroupCaptionBar = () => {
         const GroupCaptionBarCustomElement = components?.group?.CaptionBar;
@@ -22,28 +24,28 @@ const GroupElementCreationWrapper: React.FC<GroupProps> = ({ components }) => {
 
         const minimize = {
             onClick: () => {
-                webGroupsManager.minimizeGroup(options.targetId);
+                webGroupsManager.onMinimizeButtonClick(TargetType.Group, options.targetId);
             },
             ...options.minimize
         }
 
         const restore = {
             onClick: () => {
-                webGroupsManager.restoreGroup(options.targetId);
+                webGroupsManager.onRestoreButtonClick(TargetType.Group, options.targetId);
             },
             ...options.restore
         }
 
         const maximize = {
             onClick: () => {
-                webGroupsManager.maximizeGroup(options.targetId);
+                webGroupsManager.onMaximizeButtonClick(TargetType.Group, options.targetId);
             },
             ...options.maximize
         }
 
         const close = {
             onClick: () => {
-                webGroupsManager.closeGroup(options.targetId);
+                webGroupsManager.onCloseButtonClick(TargetType.Group, options.targetId);
             },
             ...options.close
         }
@@ -102,63 +104,70 @@ const GroupElementCreationWrapper: React.FC<GroupProps> = ({ components }) => {
 
             const feedback = {
                 onClick: () => {
-                    webGroupsManager.feedbackFrame(options.targetId);
+                    webGroupsManager.onFeedbackButtonClick(TargetType.Frame, options.targetId);
                 },
                 ...options.feedback
             };
 
+            const clone = {
+                onClick: () => {
+                    webGroupsManager.onCloneButtonClick(TargetType.Frame, options.targetId);
+                },
+                ...options.clone
+            };
+
             const sticky = {
                 onClick: () => {
-                    webGroupsManager.stickyFrame(options.targetId);
+                    webGroupsManager.onStickyButtonClick(TargetType.Frame, options.targetId);
                 },
                 ...options.sticky
             };
 
             const extract = {
                 onClick: () => {
-                    webGroupsManager.extractFrame(options.targetId);
+                    webGroupsManager.onExtractButtonClick(TargetType.Frame, options.targetId);
                 },
                 ...options.extract
             };
 
             const lock = {
                 onClick: () => {
-                    webGroupsManager.lockFrame(options.targetId);
+                    webGroupsManager.onLockButtonClick(TargetType.Frame, options.targetId);
                 },
                 ...options.lock
             };
 
             const unlock = {
                 onClick: () => {
-                    webGroupsManager.unlockFrame(options.targetId);
+                    webGroupsManager.onUnlockButtonClick(TargetType.Frame, options.targetId);
                 },
                 ...options.unlock
             };
 
             const minimize = {
                 onClick: () => {
-                    webGroupsManager.minimizeFrame(options.targetId);
+                    webGroupsManager.onMinimizeButtonClick(TargetType.Frame, options.targetId);
                 },
                 ...options.minimize
             }
 
             const restore = {
                 onClick: () => {
-                    webGroupsManager.restoreFrame(options.targetId);
+                    webGroupsManager.onRestoreButtonClick(TargetType.Frame, options.targetId);
                 },
                 ...options.restore
             }
 
             const maximize = {
                 onClick: () => {
-                    webGroupsManager.maximizeFrame(options.targetId);
+                    webGroupsManager.onMaximizeButtonClick(TargetType.Frame, options.targetId);
                 },
                 ...options.maximize
             }
 
             const close = {
                 onClick: () => {
-                    webGroupsManager.closeFrame(options.targetId);
+                    webGroupsManager.onCloseButtonClick(TargetType.Frame, options.targetId);
                 },
                 ...options.close
             }
@@ -208,6 +217,7 @@ const GroupElementCreationWrapper: React.FC<GroupProps> = ({ components }) => {
             return <Portal key={options.targetId} parentElement={parentElement}>
                 <FrameCaptionBarCustomElement {...options}
                     feedback={feedback}
+                    clone={clone}
                     sticky={sticky}
                     extract={extract}
                     lock={lock}
@@ -269,7 +279,7 @@ const GroupElementCreationWrapper: React.FC<GroupProps> = ({ components }) => {
     }
 
     const renderFrameLoadingAnimations = () => {
-        const LoadingAnimationCustomElement = components?.frame?.LoadingAnimation;
+        const LoadingAnimationCustomElement = LoadingAnimation;
         return Object.values(state.frameLoadingAnimations).map((wla) => {
             if (!LoadingAnimationCustomElement || !wla.parentElement) {
                 return;
@@ -402,63 +412,70 @@ const GroupElementCreationWrapper: React.FC<GroupProps> = ({ components }) => {
 
             const feedback = {
                 onClick: () => {
-                    webGroupsManager.feedbackTabBar(options.targetId);
+                    webGroupsManager.onFeedbackButtonClick(TargetType.TabBar, options.targetId);
                 },
                 ...options.feedback
             };
 
+            const clone = {
+                onClick: () => {
+                    webGroupsManager.onCloneButtonClick(TargetType.TabBar, options.targetId);
+                },
+                ...options.clone
+            };
+
             const sticky = {
                 onClick: () => {
-                    webGroupsManager.stickyTabBar(options.targetId);
+                    webGroupsManager.onStickyButtonClick(TargetType.TabBar, options.targetId);
                 },
                 ...options.sticky
             };
 
             const extract = {
                 onClick: () => {
-                    webGroupsManager.extractTabBar(options.targetId);
+                    webGroupsManager.onExtractButtonClick(TargetType.TabBar, options.targetId);
                 },
                 ...options.extract
             };
 
             const lock = {
                 onClick: () => {
-                    webGroupsManager.lockTabBar(options.targetId);
+                    webGroupsManager.onLockButtonClick(TargetType.TabBar, options.targetId);
                 },
                 ...options.lock
             }
 
             const unlock = {
                 onClick: () => {
-                    webGroupsManager.unlockTabBar(options.targetId);
+                    webGroupsManager.onUnlockButtonClick(TargetType.TabBar, options.targetId);
                 },
                 ...options.unlock
             }
 
             const minimize = {
                 onClick: () => {
-                    webGroupsManager.minimizeTabBar(options.targetId);
+                    webGroupsManager.onMinimizeButtonClick(TargetType.TabBar, options.targetId);
                 },
                 ...options.minimize
             }
 
             const restore = {
                 onClick: () => {
-                    webGroupsManager.restoreTabBar(options.targetId);
+                    webGroupsManager.onRestoreButtonClick(TargetType.TabBar, options.targetId);
                 },
                 ...options.restore
             }
 
             const maximize = {
                 onClick: () => {
-                    webGroupsManager.maximizeTabBar(options.targetId);
+                    webGroupsManager.onMaximizeButtonClick(TargetType.TabBar, options.targetId);
                 },
                 ...options.maximize
             }
 
             const close = {
                 onClick: () => {
-                    webGroupsManager.closeTabBar(options.targetId);
+                    webGroupsManager.onCloseButtonClick(TargetType.TabBar, options.targetId);
                 },
                 ...options.close
             }
@@ -478,6 +495,7 @@ const GroupElementCreationWrapper: React.FC<GroupProps> = ({ components }) => {
            
             return <Portal key={options.targetId} parentElement={parentElement}><TabButtonsCustomElement {...options}
                 feedback={feedback}
+                clone={clone}
                 sticky={sticky}
                 extract={extract}
                 lock={lock}
@@ -505,6 +523,115 @@ const GroupElementCreationWrapper: React.FC<GroupProps> = ({ components }) => {
         });
     }
 
+    const renderHtmlButtons = () => {
+        const HtmlButtonsCustomElement = components?.html?.Buttons;
+
+        return Object.values(state.htmlButtons).map((te) => {
+            if (!HtmlButtonsCustomElement || !te.parentElement) {
+                return;
+            }
+            const { parentElement, ...options } = te;
+
+            const feedback = {
+                onClick: () => {
+                    webGroupsManager.onFeedbackButtonClick(TargetType.HtmlButtons, options.targetId);
+                },
+                ...options.feedback
+            };
+
+            const clone = {
+                onClick: () => {
+                    webGroupsManager.onCloneButtonClick(TargetType.HtmlButtons, options.targetId);
+                },
+                ...options.clone
+            };
+
+            const sticky = {
+                onClick: () => {
+                    webGroupsManager.onStickyButtonClick(TargetType.HtmlButtons, options.targetId);
+                },
+                ...options.sticky
+            };
+
+            const extract = {
+                onClick: () => {
+                    webGroupsManager.onExtractButtonClick(TargetType.HtmlButtons, options.targetId);
+                },
+                ...options.extract
+            };
+
+            const lock = {
+                onClick: () => {
+                    webGroupsManager.onLockButtonClick(TargetType.HtmlButtons, options.targetId);
+                },
+                ...options.lock
+            }
+
+            const unlock = {
+                onClick: () => {
+                    webGroupsManager.onUnlockButtonClick(TargetType.HtmlButtons, options.targetId);
+                },
+                ...options.unlock
+            }
+
+            const minimize = {
+                onClick: () => {
+                    webGroupsManager.onMinimizeButtonClick(TargetType.HtmlButtons, options.targetId);
+                },
+                ...options.minimize
+            }
+
+            const restore = {
+                onClick: () => {
+                    webGroupsManager.onRestoreButtonClick(TargetType.HtmlButtons, options.targetId);
+                },
+                ...options.restore
+            }
+
+            const maximize = {
+                onClick: () => {
+                    webGroupsManager.onMaximizeButtonClick(TargetType.HtmlButtons, options.targetId);
+                },
+                ...options.maximize
+            }
+
+            const close = {
+                onClick: () => {
+                    webGroupsManager.onCloseButtonClick(TargetType.HtmlButtons, options.targetId);
+                },
+                ...options.close
+            }
+
+            let customButtonsProps = new Array<CustomButtonProps>();
+            if (options.customButtons) {
+                customButtonsProps = options.customButtons.map((cButton) => { 
+                    return {
+                        onClick: () => {
+                            webGroupsManager.clickCustomButton(options.targetId, cButton.buttonId);
+                        },
+                        visible: true,
+                        imageData: cButton.imageData,
+                        tooltip: cButton.tooltip,
+                        buttonId: cButton.buttonId}});
+            }
+           
+            return <Portal key={options.targetId} parentElement={parentElement}><HtmlButtonsCustomElement {...options}
+                feedback={feedback}
+                clone={clone}
+                sticky={sticky}
+                extract={extract}
+                lock={lock}
+                unlock={unlock}
+                minimize={minimize}
+                maximize={maximize}
+                restore={restore}
+                close={close}
+                customButtons={customButtonsProps}
+                frameId={options.targetId}
+            /></Portal>
+        });
+    }
+
     return <>
         {renderGroupCaptionBar()}
         {renderGroupOverlay()}
@@ -520,6 +647,7 @@ const GroupElementCreationWrapper: React.FC<GroupProps> = ({ components }) => {
         {renderAfterTabsZones()}
         {renderTabHeaderButtons()}
         {renderBelowTabs()}
+        {renderHtmlButtons()}
         <GroupWrapper
             onCreateGroupCaptionBarRequested={components?.group?.CaptionBar ? webGroupsStore.onCreateGroupCaptionBarRequested : undefined}
             onCreateGroupOverlayRequested={components?.group?.Overlay ? webGroupsStore.onCreateGroupOverlayRequested : undefined}
@@ -527,7 +655,7 @@ const GroupElementCreationWrapper: React.FC<GroupProps> = ({ components }) => {
             onCreateFrameWindowOverlayRequested={components?.frame?.Overlay ? webGroupsStore.onCreateFrameWindowOverlayRequested : undefined}
             onCreateAboveWindowRequested={components?.frame?.AboveWindow ? webGroupsStore.onCreateAboveWindowRequested : undefined}
             onCreateWindowContentOverlayRequested={components?.frame?.WindowContentOverlay ? webGroupsStore.onCreateWindowContentOverlayRequested : undefined}
-            onCreateFrameLoadingAnimationRequested={components?.frame?.LoadingAnimation ? webGroupsStore.onCreateFrameLoadingAnimationRequested : undefined}
+            onCreateFrameLoadingAnimationRequested={LoadingAnimation ? webGroupsStore.onCreateFrameLoadingAnimationRequested : undefined}
             onCreateBelowWindowRequested={components?.frame?.BelowWindow ? webGroupsStore.onCreateBelowWindowRequested : undefined}
             onCreateAboveTabsRequested={components?.tabs?.Above ? webGroupsStore.onCreateAboveTabsComponentRequested : undefined}
             onCreateBeforeTabsComponentRequested={components?.tabs?.Before ? webGroupsStore.onCreateBeforeTabsComponentRequested : undefined}
@@ -535,6 +663,7 @@ const GroupElementCreationWrapper: React.FC<GroupProps> = ({ components }) => {
             onCreateAfterTabsComponentRequested={components?.tabs?.After ? webGroupsStore.onCreateAfterTabsComponentRequested : undefined}
             onCreateTabHeaderButtonsRequested={components?.tabs?.Buttons ? webGroupsStore.onCreateTabHeaderButtonsRequested : undefined}
             onCreateBelowTabsRequested={components?.tabs?.Below ? webGroupsStore.onCreateBelowTabsComponentRequested : undefined}
+            onCreateHtmlButtonsRequested={components?.html?.Buttons ? webGroupsStore.onCreateHtmlButtonsRequested : undefined}
             onUpdateGroupCaptionBarRequested={components?.group?.CaptionBar ? webGroupsStore.onUpdateGroupCaptionBarRequested : undefined}
             onUpdateFrameWindowOverlayRequested={components?.group?.Overlay ? webGroupsStore.onUpdateFrameWindowOverlayRequested : undefined}
             onUpdateFrameCaptionBarRequested={components?.flat?.CaptionBar ? webGroupsStore.onUpdateFrameCaptionBarRequested : undefined}
@@ -558,6 +687,7 @@ const GroupElementCreationWrapper: React.FC<GroupProps> = ({ components }) => {
             onRemoveAfterTabsComponentRequested={webGroupsStore.onRemoveAfterTabsComponentRequested}
             onRemoveTabHeaderButtonsRequested={webGroupsStore.onRemoveTabHeaderButtonsRequested}
             onRemoveBelowTabsRequested={webGroupsStore.onRemoveBelowTabsRequested}
+            onRemoveHtmlButtonsRequested={webGroupsStore.onRemoveHtmlButtonsRequested}
             onShowCaptionEditorRequested={webGroupsStore.onShowCaptionEditorRequested}
             onCommitCaptionEditingRequested={webGroupsStore.onCommitCaptionEditingRequested}
             onHideCaptionEditorRequested={webGroupsStore.onHideCaptionEditorRequested}