+++ changelog.md
@@ -1,3 +1,13 @@
+2.0.6
+fix: flat channels selector fix
+2.0.4
+fix: typo in loading animation class name
+2.0.3
+chore: release process issue
+2.0.2
+fix: started using the default loading animation component
+2.0.1
+feat: customizable html buttons
 2.0.0
 feat: rebranding
 1.4.0